Fixes for runner acceptance test to pass
features/support/kalibro_cucumber_helpers.yml should be ignored and left a sample
KalibroGatekeeper client updated to fix Metric's compound method to return a boolean
AnalizoMetricCollector should persist the KalibroModule
Overriding getters and setters for ActiveRecord attributes is not a good idea. So renamed granularity to granlrty and name to long_name.