Adds a default value to repository branch.
Just adding the branch column would make already existing repositorys to remain
with a NULL branch which will then break their processing.

Adding master as the default branch is a workaround since master may not be the
default branch for such repositories. As well adding a not null restriction
will enforce that the database add the default value to already existing
entries.