Added a call to destroy the previous process times on preparer
Also refactored the processing factory to ensure that the process times
are created on each test that actually needs them. Additionally, stubbed
the process times on the preparer unit test to ensure that the call to
destroy_all is being made on the stubbed process time list, instead of
a call on an actual active record association.

Signed-off-by: Daniel Miranda <danielkza2@gmail.com>