Fixed availability check of the Radon collector
Previously it was completely broket by simply checking if a string is
nil instead of using backticks to call the command. But even that would
be wrong, since it raises an exception if the executable is not found,
and never returns nil.

Implement a method in the base collector to handle that case and use it.
It should probably be applied to other collectors as well later.