Updated HotspotResult database and model with new architecture
It uses a table of RelatedHotspotResults that just represents an
aggregation of results. It has no data, so it just has an id,
and each HotspotResult has the foreign key connected to it.

The model then uses a has_many :through association to find the related
results directly.

Fixing the Flay collector is still pending.

Signed off by: Pedro Scocco <pedroscocco@gmail.com>