Fixes the calculation of metrics with SOFTWARE scope
Analizo created a module result with SOFTWARE scope and named with an
empty string. The TreeBuilder state tries to create a root module result
with name "ROOT", if it can't find it on the database.
That, along with the aggregation of two metric results with SOFTWARE granularities, was
leading to inconsistencies that made Kalibro lose information about
results of analizo metrics with SOFTWARE scope.

Now, the analizo parser tries to create a new module result already with
name "ROOT". It avoids the creation of new root module results and
prevents the aggregation of two metric results with SOFTWARE
granularities (which does not make sense, anyway).

Signed off by: Daniel Miranda <danielkza2@gmail.com>