Replace Analizo collector with Kolekti gem
The Kolekti gem will abstract common code between collectors, that in
the future will each be their own gem. For now. replace only the Analizo
collector as the first prototype collector to be ported.

Kolekti does things just a little bit differently than we previously
did. Instead of exposing which collector is responsible for processing
which metric, it simply accepts a list of desired metric for collection,
and distributes them among the collectors internally.

It also decouples the production of metrics from their persistence: the
collector will merely call-back to the client application, which will
then decide how to save the results. A persistence strategy was
implemented based on code previously repeated in all the collectors. It
should actually work more reliably than before, by being shared among
different collectors, and using proper transactions to group creation of
multiple entities that should not exist without each other.

To keep the unported collectors working, special logic had to be added
to the Collector processing step, to distribute wanted metrics between
Kolekti and other pre-existing collectors. Unification should be
possible when all the collectors are porting, and no collector selection
logic should be left (it should be completely delegated to Kolekti).

Kolekti also requires collectors to be manually enable as of now. Add an
initializer in the configuration enabling the Analizo collector, which
is the only one available at the moment (and built-in into the main
Kolekti gem).

An acceptance test using Analizo was re-added (after being removed due
to slowness a bit ago). It may be wise to keep it, but just leave it
behind a tag in the future if it still proves problematic.

Signed off by: Heitor Reis <marcheing@gmail.com>