Create javascript functions for hotspot metrics
Until now, we were not creating them. However, now we want to raise more
meaningful errors to our users.

The error here is that we do not want to let the user create compound
metric scripts that use hotspot metric codes. That is invalid because
hotspot metric results should return messages and chunks of code, not
values that can be used to calculate a compound metric.

The solution was to add a piece of code on the CompoundResultCalculator
that creates functions signed by the code of the hotspot metric (just as
we do with tree metrics) and define the body as a javscript that throws
an error with a meaningful error message for this case.

Notice that the error will only be raised if there's actually a compound
metric that uses a hotspot metric code on its script. Therefore,
compound metrics that do not use hotspot metric codes should be fine.
This solution makes the processing finish with an error state if there's
a single broken compound metric.