Update database_cleaner setup to use truncation when needed
The transaction strategy was only working because we didn't have
referential integrity checks. The job workers are not guaranteed to stay
in the same the same transaction as the tests - the PHPMD tests, for
example, spawned a new thread - and therefore, we must use truncation
instead so all the models become visible to the workers in time.