Add indexes to foreign key references that are not primary keys
PostgreSQL does not create them automatically, but we want them when
deletions are cascade to other tables - otherwise we'll have to scan the
whole table at that time. Some of the are also used for searching, such
as the processing_id in process_times, or the project_id in
repositories.